platform: make cluster.Machines() less surprising
Change cluster.Machines() to be less surprising. Since it returns a
slice, kola test-writers may assume that the order of this is always the
same.  While changing the backing structure from a map to a slice still
doesn't make this strictly true because elements can be deleted, it will
make it true in the common case. Normally, machines are not torn down
until the end of a test.

I think this is worthwhile regardless if the better advice is to always
carry pointers to machines created in a test.