kola/docker: test docker skim
This PR reworks the docker tests in kola such that any test may be
targetted at a combination of target configurations.

The current configurations are 'host' and 'skim-1.12.6', but it should
be easy to add more.

This effectively doubles the number of docker tests and thus does add
not-insignificant runtime.