kola/harness: add nondestructive flag for tests
Adds a NonDestructive flag for tests similar to Min/MaxVersion
flags which allows a test to denote that it is safe to run on the
same machine as other non-destructive tests. Currently for a test
to properly be non-destructive they must leave the machine in the
exact state that they received it in, even in the case of an error
during the test.

These tests can still be filtered down as usual, they will just be
ran on a cluster that is shared with any other tests which require
the same cluster setup.

In the `_kola_tmp` output a symlink is created for each test which
points to the output directory for the cluster that it ran on.