platform/aws: automatically resolve dupe images
This is intended to automatically resolve duplicate image name errors
which frequently crop up during pre-release.

This change does not stop 'plume release' from failing if there
are multiple images. This is an error scenario we may wish to handle as
well, in which case `FindImage` can be removed entirely.
For now, I felt it safer to limit the changes to the pre-release
operations.