cmd/ore/oci: update OCI ore commands for new SDK
There were a few changes in the returned structs between the two SDK's.
This updates `create-kola-vcn` & `delete-kola-vcn` to use the correct
fields/types for the new structs.

`delete-kola-vcn` has been disabled due to instability in the command,
the networking objects fail to be deleted as they seem to depend on
each other in a cyclical fashion which requires multiple runs of the
command to actually perform a successful delete.